LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-003-expected.txt


FAIL Transition with anchor names defined in three different tree scopes assert_equals: expected 300 but got 0
